+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [4.0.0] - 2022-01-10
+### Changed
+- Updated to https://github.com/pronamic/wp-pay-core/releases/tag/4.0.0.
+- Added support for checkboxes field type.
+- Don't force description to have a dynamic part (use `[id]` in transaction description setting instead).
+
 ## [3.0.0] - 2021-08-05
 - Updated to `pronamic/wp-pay-core`  version `3.0.0`.
 - Updated to `pronamic/wp-money`  version `2.0.0`.
@@ -66,7 +72,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-11-05
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/formidable-forms/compare/3.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/formidable-forms/compare/4.0.0...HEAD
+[4.0.0]: https://github.com/wp-pay-extensions/formidable-forms/compare/3.0.0...4.0.0
 [3.0.0]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.2.1...3.0.0
 [2.2.1]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.2.0...2.2.1
 [2.2.0]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.1.4...2.2.0